requirements.txt
ollama
streamlit
pymupdf      
requests
sendgrid
schedule
python-dotenv
openai    
langchain
google-generativeai==0.8.1